refactor: rename booleen into bool
