+++ csv_detective/formats/bool.py
